Adaptive P-value Thresholding with Generalized Linear Models

adapt_glm is a wrapper of adapt that fits pi(x) and mu(x) by glm.

adapt_glm(
  x,
  pvals,
  pi_formulas,
  mu_formulas,
  dist = beta_family(),
  s0 = rep(0.45, length(pvals)),
  alphas = seq(0.01, 1, 0.01),
  piargs = list(),
  muargs = list(),
  ...
)

Arguments

x
covariates (i.e. side-information). Should be compatible to models. See Details

pvals
a vector of values in [0, 1]. P-values

pi_formulas
a vector/list of strings/formulas. Formulas for fitting pi(x) by glm. See Details

mu_formulas
a vector/list of strings/formulas. Formulas for fitting mu(x) by glm. See Details

dist
an object of class "gen_exp_family". beta_family() as default

s0
a vector of values in [0, 0.5). Initial threshold.

alphas
a vector of values in (0, 1). Target FDR levels.

piargs
a list. Other arguments passed to glm for fitting pi(x)

muargs
a list. Other arguments passed to glm for fitting mu(x)

...
other arguments passed to adapt (except models)

Details

pi_formulas and mu_formulas can either be a list or a vector with each element being a string or a formula. For instance, suppose x has a single column with name x1, the following five options are valid for the same inputs (ns forms a spline basis with df knots):

c("x1", "ns(x1, df = 8)");

c("~ x1", "~ ns(x1, df = 8)");

list("x1", "ns(x1, df = 8)");

list("~ x1", "~ ns(x1, df = 8)");

list(~ x1, ~ ns(x1, df = 8))

There is no need to specify the name of the response variable, as this is handled in the function.

When x has a few variables, it is common to use non-parametric GLM by replacing x by a spline basis of x. In this case, ns from library(splines) package is suggested.

See also

Examples

# \donttest{
# Load estrogen data
data(estrogen)
pvals <- as.numeric(estrogen$pvals)
x <- data.frame(x = as.numeric(estrogen$ord_high))
dist <- beta_family()

# Subsample the data for convenience
inds <- (x$x <= 5000)
pvals <- pvals[inds]
x <- x[inds,,drop = FALSE]

# Run adapt_glm
library("splines")
formulas <- paste0("ns(x, df = ", 6:10, ")")
res <- adapt_glm(x = x, pvals = pvals, pi_formulas = formulas,
                 mu_formulas = formulas, dist = dist, nfits = 10)

#> Model selection starts!
#> Shrink the set of candidate models if it is too time-consuming.
#>   |                                                          |                                                  |   0%  |                                                          |==========                                        |  20%  |                                                          |====================                              |  40%  |                                                          |==============================                    |  60%  |                                                          |========================================          |  80%  |                                                          |==================================================| 100%
#> alpha = 0.29: FDPhat 0.2899, Number of Rej. 3474
#> alpha = 0.28: FDPhat 0.2798, Number of Rej. 3306
#> alpha = 0.27: FDPhat 0.27, Number of Rej. 3182
#> alpha = 0.26: FDPhat 0.26, Number of Rej. 2904
#> alpha = 0.25: FDPhat 0.2498, Number of Rej. 2778
#> alpha = 0.24: FDPhat 0.2399, Number of Rej. 2664
#> alpha = 0.23: FDPhat 0.2297, Number of Rej. 2608
#> alpha = 0.22: FDPhat 0.2198, Number of Rej. 2552
#> alpha = 0.21: FDPhat 0.2098, Number of Rej. 2469
#> alpha = 0.2: FDPhat 0.1999, Number of Rej. 2376
#> alpha = 0.19: FDPhat 0.1898, Number of Rej. 2292
#> alpha = 0.18: FDPhat 0.1799, Number of Rej. 2173
#> alpha = 0.17: FDPhat 0.1698, Number of Rej. 2096
#> alpha = 0.16: FDPhat 0.1598, Number of Rej. 2040
#> alpha = 0.15: FDPhat 0.15, Number of Rej. 1980
#> alpha = 0.14: FDPhat 0.1399, Number of Rej. 1915
#> alpha = 0.13: FDPhat 0.1296, Number of Rej. 1844
#> alpha = 0.12: FDPhat 0.1198, Number of Rej. 1778
#> alpha = 0.11: FDPhat 0.1096, Number of Rej. 1715
#> alpha = 0.1: FDPhat 0.0999, Number of Rej. 1482
#> alpha = 0.09: FDPhat 0.0899, Number of Rej. 1390
#> alpha = 0.08: FDPhat 0.0798, Number of Rej. 1303
#> alpha = 0.07: FDPhat 0.0693, Number of Rej. 1155
#> alpha = 0.06: FDPhat 0.06, Number of Rej. 984
#> alpha = 0.05: FDPhat 0.0497, Number of Rej. 925
#> alpha = 0.04: FDPhat 0.0393, Number of Rej. 839
#> alpha = 0.03: FDPhat 0.0289, Number of Rej. 589
#> alpha = 0.02: FDPhat 0.019, Number of Rej. 263

# Run adapt by manually setting models for glm
models <- lapply(formulas, function(formula){
    piargs <- muargs <- list(formula = formula)
    gen_adapt_model(name = "glm", piargs = piargs, muargs = muargs)
})
res2 <- adapt(x = x, pvals = pvals, models = models,
              dist = dist, nfits = 10)

#> Model selection starts!
#> Shrink the set of candidate models if it is too time-consuming.
#>   |                                                          |                                                  |   0%  |                                                          |==========                                        |  20%  |                                                          |====================                              |  40%  |                                                          |==============================                    |  60%  |                                                          |========================================          |  80%  |                                                          |==================================================| 100%
#> alpha = 0.29: FDPhat 0.2899, Number of Rej. 3474
#> alpha = 0.28: FDPhat 0.2798, Number of Rej. 3306
#> alpha = 0.27: FDPhat 0.27, Number of Rej. 3182
#> alpha = 0.26: FDPhat 0.26, Number of Rej. 2904
#> alpha = 0.25: FDPhat 0.2498, Number of Rej. 2778
#> alpha = 0.24: FDPhat 0.2399, Number of Rej. 2664
#> alpha = 0.23: FDPhat 0.2297, Number of Rej. 2608
#> alpha = 0.22: FDPhat 0.2198, Number of Rej. 2552
#> alpha = 0.21: FDPhat 0.2098, Number of Rej. 2469
#> alpha = 0.2: FDPhat 0.1999, Number of Rej. 2376
#> alpha = 0.19: FDPhat 0.1898, Number of Rej. 2292
#> alpha = 0.18: FDPhat 0.1799, Number of Rej. 2173
#> alpha = 0.17: FDPhat 0.1698, Number of Rej. 2096
#> alpha = 0.16: FDPhat 0.1598, Number of Rej. 2040
#> alpha = 0.15: FDPhat 0.15, Number of Rej. 1980
#> alpha = 0.14: FDPhat 0.1399, Number of Rej. 1915
#> alpha = 0.13: FDPhat 0.1296, Number of Rej. 1844
#> alpha = 0.12: FDPhat 0.1198, Number of Rej. 1778
#> alpha = 0.11: FDPhat 0.1096, Number of Rej. 1715
#> alpha = 0.1: FDPhat 0.0999, Number of Rej. 1482
#> alpha = 0.09: FDPhat 0.0899, Number of Rej. 1390
#> alpha = 0.08: FDPhat 0.0798, Number of Rej. 1303
#> alpha = 0.07: FDPhat 0.0693, Number of Rej. 1155
#> alpha = 0.06: FDPhat 0.06, Number of Rej. 984
#> alpha = 0.05: FDPhat 0.0497, Number of Rej. 925
#> alpha = 0.04: FDPhat 0.0393, Number of Rej. 839
#> alpha = 0.03: FDPhat 0.0289, Number of Rej. 589
#> alpha = 0.02: FDPhat 0.019, Number of Rej. 263

# Check equivalence
identical(res, res2)

#> [1] TRUE

# }