Adaptive P-value Thresholding with L1/L2 Penalized Generalized Linear Models

adapt_glmnet is a wrapper of adapt that fits pi(x) and mu(x) by glmnet from glmnet package.

adapt_glmnet(
  x,
  pvals,
  piargs = list(),
  muargs = list(),
  dist = beta_family(),
  s0 = rep(0.45, length(pvals)),
  alphas = seq(0.01, 1, 0.01),
  ...
)

Arguments

x
covariates (i.e. side-information). Should be compatible to models. See Details

pvals
a vector of values in [0, 1]. P-values

piargs
a list. Other arguments passed to glmnet for fitting pi(x)

muargs
a list. Other arguments passed to glmnet for fitting mu(x)

dist
an object of class "gen_exp_family". beta_family() as default

s0
a vector of values in [0, 0.5). Initial threshold.

alphas
a vector of values in (0, 1). Target FDR levels.

...
other arguments passed to adapt (except models)

Details

adapt_glmnet by default implements LASSO on x with lambda selected by cross-validation. Specify in piargs and muargs if ridge or elastic-net penalty is needed.

See also

Examples

# \donttest{
# Generate a 100-dim covariate x
set.seed(0)
m <- 100
n <- 1000
x <- matrix(runif(n * m), n, m)

# Generate the parameters from a conditional two-group
# logistic-Gamma GLM  where pi(x) and mu(x) are both
# linear in x. pi(x) has an intercept so that the average
# of pi(x) is 0.3
inv_logit <- function(x) {exp(x) / (1 + exp(x))}
pi1 <- 0.3
beta.pi <- c(3, 3, rep(0, m-2))
beta0.pi <- uniroot(function(b){
    mean(inv_logit(x %*% beta.pi + b)) - pi1
}, c(-100, 100))$root
pi <- inv_logit(x %*% beta.pi + beta0.pi)
beta.mu <- c(2, 2, rep(0, m-2))
beta0.mu <- 0
mu <- pmax(1, x %*% beta.mu + beta0.mu)

# Generate p-values
H0 <- as.logical(ifelse(runif(n) < pi, 1, 0))
y <- ifelse(H0, rexp(n, 1/mu), rexp(n, 1))
pvals <- exp(-y)

# Run adapt_glmnet
res <- adapt_glmnet(x, pvals, s0 = rep(0.15, n), nfits = 5)

#> alpha = 0.65: FDPhat 0.6471, Number of Rej. 221
#> alpha = 0.64: FDPhat 0.64, Number of Rej. 200
#> alpha = 0.63: FDPhat 0.6294, Number of Rej. 197
#> alpha = 0.62: FDPhat 0.6193, Number of Rej. 197
#> alpha = 0.61: FDPhat 0.6082, Number of Rej. 194
#> alpha = 0.6: FDPhat 0.5959, Number of Rej. 193
#> alpha = 0.59: FDPhat 0.5885, Number of Rej. 192
#> alpha = 0.58: FDPhat 0.5781, Number of Rej. 192
#> alpha = 0.57: FDPhat 0.5677, Number of Rej. 192
#> alpha = 0.56: FDPhat 0.5585, Number of Rej. 188
#> alpha = 0.55: FDPhat 0.5484, Number of Rej. 186
#> alpha = 0.54: FDPhat 0.5351, Number of Rej. 185
#> alpha = 0.53: FDPhat 0.5297, Number of Rej. 185
#> alpha = 0.52: FDPhat 0.5189, Number of Rej. 185
#> alpha = 0.51: FDPhat 0.506, Number of Rej. 166
#> alpha = 0.5: FDPhat 0.5, Number of Rej. 166
#> alpha = 0.49: FDPhat 0.4839, Number of Rej. 155
#> alpha = 0.48: FDPhat 0.4768, Number of Rej. 151
#> alpha = 0.47: FDPhat 0.4667, Number of Rej. 150
#> alpha = 0.46: FDPhat 0.4564, Number of Rej. 149
#> alpha = 0.45: FDPhat 0.4497, Number of Rej. 149
#> alpha = 0.44: FDPhat 0.4384, Number of Rej. 146
#> alpha = 0.43: FDPhat 0.4247, Number of Rej. 146
#> alpha = 0.42: FDPhat 0.4138, Number of Rej. 145
#> alpha = 0.41: FDPhat 0.4069, Number of Rej. 145
#> alpha = 0.4: FDPhat 0.4, Number of Rej. 145
#> alpha = 0.39: FDPhat 0.3846, Number of Rej. 143
#> alpha = 0.38: FDPhat 0.3786, Number of Rej. 140
#> alpha = 0.37: FDPhat 0.3657, Number of Rej. 134
#> alpha = 0.36: FDPhat 0.3582, Number of Rej. 134
#> alpha = 0.35: FDPhat 0.3459, Number of Rej. 133
#> alpha = 0.34: FDPhat 0.3333, Number of Rej. 129
#> alpha = 0.33: FDPhat 0.3254, Number of Rej. 126
#> alpha = 0.32: FDPhat 0.3175, Number of Rej. 126
#> alpha = 0.31: FDPhat 0.3095, Number of Rej. 126
#> alpha = 0.3: FDPhat 0.2975, Number of Rej. 121
#> alpha = 0.29: FDPhat 0.2857, Number of Rej. 119
#> alpha = 0.28: FDPhat 0.2773, Number of Rej. 119
#> alpha = 0.27: FDPhat 0.2689, Number of Rej. 119
#> alpha = 0.26: FDPhat 0.2521, Number of Rej. 119
#> alpha = 0.25: FDPhat 0.2458, Number of Rej. 118
#> alpha = 0.24: FDPhat 0.2373, Number of Rej. 118
#> alpha = 0.23: FDPhat 0.2288, Number of Rej. 118
#> alpha = 0.22: FDPhat 0.2155, Number of Rej. 116
#> alpha = 0.21: FDPhat 0.2039, Number of Rej. 103
#> alpha = 0.2: FDPhat 0.1942, Number of Rej. 103
#> alpha = 0.19: FDPhat 0.1856, Number of Rej. 97
#> alpha = 0.18: FDPhat 0.1771, Number of Rej. 96
#> alpha = 0.17: FDPhat 0.1684, Number of Rej. 95
#> alpha = 0.16: FDPhat 0.1481, Number of Rej. 81
#> alpha = 0.15: FDPhat 0.1481, Number of Rej. 81
#> alpha = 0.14: FDPhat 0.1375, Number of Rej. 80
#> alpha = 0.13: FDPhat 0.129, Number of Rej. 62
#> alpha = 0.12: FDPhat 0.1129, Number of Rej. 62
#> alpha = 0.11: FDPhat 0.0968, Number of Rej. 62
#> alpha = 0.1: FDPhat 0.0968, Number of Rej. 62
#> alpha = 0.09: FDPhat 0.0862, Number of Rej. 58
#> alpha = 0.08: FDPhat 0.0741, Number of Rej. 54
#> alpha = 0.07: FDPhat 0.0556, Number of Rej. 54
#> alpha = 0.06: FDPhat 0.0556, Number of Rej. 54
#> alpha = 0.05: FDPhat 0.0392, Number of Rej. 51
#> alpha = 0.04: FDPhat 0.0392, Number of Rej. 51
#> alpha = 0.03: FDPhat 0.0213, Number of Rej. 47

# }