Quantifying Information Loss of Adaptive P-Value Thresholding

corr_lfdr computes the oracle local FDR estimate, by using revealing all p-values, and computes the Pearson correlation between it and the estimate within each step of adapt.

corr_lfdr(obj, x, pvals, model = NULL, niter_oracle = 100)

Arguments

obj
an 'adapt' object. Output of adapt function

x
covariates (i.e. side-information). Should be compatible to models.

pvals
a vector of values in [0, 1]. P-values

model
an optional argument. If model = NULL then the last model in obj$models is used for fitting the oracle model (i.e. with all p-values revealed). Otherwise it should be an 'adapt_model' object

niter_oracle
an positive integer. Number of iterations in EM algorithm

Value

corra vector of values in [0, 1]. Pearson correlation of oracle local FDR estimate and the estimates within each step. Each value corresponds to an entry of obj$params

oracle_lfdra vector of values in [0, 1]. Oracle local FDR estimate

lfdra matrix of values in [0, 1]. Local FDR estimates within each step.

alphasa vector of values in [0, 1]. The target FDR levels corresponding to each local FDR estimate

nmasksa vector of integers. The number of masked p-values corresponding to each local FDR estimate

Examples

# \donttest{
# Load estrogen data
data(estrogen)
pvals <- as.numeric(estrogen$pvals)
x <- data.frame(x = as.numeric(estrogen$ord_high))
dist <- beta_family()

# Subsample the data for convenience
inds <- (x$x <= 5000)
pvals <- pvals[inds]
x <- x[inds,,drop = FALSE]

# Run adapt_glm
library("splines")
formulas <- paste0("ns(x, df = ", 6:10, ")")
res <- adapt_glm(x = x, pvals = pvals, pi_formulas = formulas,
                 mu_formulas = formulas, dist = dist, nfits = 10)

#> Model selection starts!
#> Shrink the set of candidate models if it is too time-consuming.
#>   |                                                          |                                                  |   0%  |                                                          |==========                                        |  20%  |                                                          |====================                              |  40%  |                                                          |==============================                    |  60%  |                                                          |========================================          |  80%  |                                                          |==================================================| 100%
#> alpha = 0.29: FDPhat 0.2899, Number of Rej. 3474
#> alpha = 0.28: FDPhat 0.2798, Number of Rej. 3306
#> alpha = 0.27: FDPhat 0.27, Number of Rej. 3182
#> alpha = 0.26: FDPhat 0.26, Number of Rej. 2904
#> alpha = 0.25: FDPhat 0.2498, Number of Rej. 2778
#> alpha = 0.24: FDPhat 0.2399, Number of Rej. 2664
#> alpha = 0.23: FDPhat 0.2297, Number of Rej. 2608
#> alpha = 0.22: FDPhat 0.2198, Number of Rej. 2552
#> alpha = 0.21: FDPhat 0.2098, Number of Rej. 2469
#> alpha = 0.2: FDPhat 0.1999, Number of Rej. 2376
#> alpha = 0.19: FDPhat 0.1898, Number of Rej. 2292
#> alpha = 0.18: FDPhat 0.1799, Number of Rej. 2173
#> alpha = 0.17: FDPhat 0.1698, Number of Rej. 2096
#> alpha = 0.16: FDPhat 0.1598, Number of Rej. 2040
#> alpha = 0.15: FDPhat 0.15, Number of Rej. 1980
#> alpha = 0.14: FDPhat 0.1399, Number of Rej. 1915
#> alpha = 0.13: FDPhat 0.1296, Number of Rej. 1844
#> alpha = 0.12: FDPhat 0.1198, Number of Rej. 1778
#> alpha = 0.11: FDPhat 0.1096, Number of Rej. 1715
#> alpha = 0.1: FDPhat 0.0999, Number of Rej. 1482
#> alpha = 0.09: FDPhat 0.0899, Number of Rej. 1390
#> alpha = 0.08: FDPhat 0.0798, Number of Rej. 1303
#> alpha = 0.07: FDPhat 0.0693, Number of Rej. 1155
#> alpha = 0.06: FDPhat 0.06, Number of Rej. 984
#> alpha = 0.05: FDPhat 0.0497, Number of Rej. 925
#> alpha = 0.04: FDPhat 0.0393, Number of Rej. 839
#> alpha = 0.03: FDPhat 0.0289, Number of Rej. 589
#> alpha = 0.02: FDPhat 0.019, Number of Rej. 263

# Run corr_lfdr
obj <- corr_lfdr(res, x, pvals)
obj$corr

#>  [1] 0.7792830 0.8076822 0.8544305 0.9168807 0.9587528 0.9792674 0.9903270
#>  [8] 0.9960151 0.9984551 0.9990748

# }