Generate exp_family Objects for Exponential Families

exp_family objects contain all required information in an exponential family to perform the E-step. The exponential function is encoded by $$h(p; \mu) = \exp\{(\eta(\mu) - \eta(\mu^{*})) g(p) - (A(\mu) - A(\mu^{*}))\}$$ where \(g(p)\) is an arbitrary transformation, \(\mu\) is the mean parameter, \(\eta\) is the natural parameter, and \(A(\mu)\) is the partition function. The extra redundant parameter \(\mu^{*}\) is to guarantee that \(U([0, 1])\) belongs to the class.

gen_exp_family(g, ginv, eta, mustar, A, name = NULL, family = NULL)

beta_family()

inv_gaussian_family()

Arguments

g
a function. An transformation of p-values

ginv
a function. The inverse function of g

eta
a function. The natural parameter as a function of the mean parameter mu

mustar
a scalar. The mean parameter that gives \(U([0, 1])\)

A
a function. The partition function

name
a string. A name for the family. NULL by default

family
an object of class "family" from stats package. The family used for model fitting in glm, gam, glmnet, etc..

Value

an object of class "exp_family". This includes all inputs and h, the density function.

Details

Beta family (beta_family()): modeling p-values as Beta-distributed random variables, i.e. \(g(p) = -log(p)\), \(\eta(\mu) = -1 / \mu\), \(\mu* = 1\), \(A(\mu) = log(\mu)\), name = "beta" and family = Gamma(). Beta-family is highly recommended for general problems and used as default.

Inverse-gaussian family (inv_gaussian_family()): modeling p-values as transformed z-scores, i.e. \(g(p) = \Phi^{-1}(p) (\Phi is the c.d.f. of a standard normal random variable)\), \(\eta(\mu) = \mu\), \(\mu* = 0\), \(A(\mu) = \mu^2 / 2\), name = "inv_gaussian" and family = gaussian().